Initial commit
This is a hackernews app created for learing elmish purposes. This is code is mostly following the Zaid-Ajaj elmish book steps with some of the functionalities implemented  by me as exercise. The original code comes from this link https://github.com/Zaid-Ajaj/elmish-hackernews-part3. Please do not hesitate to leave a comment if you have any.